refactor: dropdown menu to components
