+++ site/src/components/station_dropdown_menu/index.tsx
@@ -1,11 +1,12 @@
 import type { Locale } from '~/types/common'
 
-import React from 'react'
+import { DialogTrigger } from '@radix-ui/react-dialog'
 import dynamic from 'next/dynamic'
 import Link from 'next/link'
-import { DialogTrigger } from '@radix-ui/react-dialog'
+import React from 'react'
 import { shallow } from 'zustand/shallow'
 
+import { CheckboxItem, DropdownMenu, Item } from '~/components/dropdown_menu'
 import CirclesHorizontal from '~/components/icons/circles_horizontal.svg'
 import Close from '~/components/icons/close.svg'
 import Filter from '~/components/icons/filter.svg'
@@ -14,7 +15,6 @@ import HeartFilled from '~/components/icons/heart_filled.svg'
 import HeartOutline from '~/components/icons/heart_outline.svg'
 import ToBottom from '~/components/icons/to_bottom.svg'
 import ToTop from '~/components/icons/to_top.svg'
-import { CheckboxItem, DropdownMenu, Item } from '~/features/dropdown_menu'
 import { useFavorites } from '~/hooks/use_favorites'
 import { useStationFilters } from '~/hooks/use_filters'
 import { useStationPage } from '~/hooks/use_station_page'
