--- site/src/features/dropdown_menu/index.ts
@@ -1,4 +1,0 @@
-export * from './components/dropdown_menu'
-export * from './components/checkbox_item'
-export * from './components/item'
-export * from './components/icon'
